Refactor quantitative model to import logging and configuration management from separate modules.
